Separate tests from the build script
