+++ build.js
@@ -1,33 +1,15 @@
-
 var fs = require('fs')
-var charmap = require('./charmap')
 
-
-exports.sort = function () {
+function sortCharmap () {
+  var charmap = JSON.parse(fs.readFileSync('charmap.json', 'utf8'))
   var sorted = charmap.sort(function (a, b) {
     return a.key > b.key ? 1 : a.key < b.key ? -1 : 0
   })
   fs.writeFileSync('charmap.json', JSON.stringify(sorted, null, 2), 'utf8')
 }
 
-exports.duplicates = function () {
-  var cache = {}
-  var duplicates = []
-  charmap.forEach(function (pair) {
-    if (!cache[pair.key]) {
-      cache[pair.key] = pair
-    }
-    else {
-      duplicates.push(pair)
-    }
-  })
-  return {
-    cache: cache,
-    duplicates: duplicates
-  }
-}
-
-exports.replace = function () {
+function replaceCharmap () {
+  var charmap = JSON.parse(fs.readFileSync('charmap.json', 'utf8'))
   var obj = charmap.reduce(function (obj, pair) {
     obj[pair.key] = pair.value
     return obj
@@ -44,3 +26,6 @@ exports.replace = function () {
 
   fs.writeFileSync('index.js', source, 'utf8')
 }
+
+sortCharmap()
+replaceCharmap()
